+++ aim/deploy/docker/server_prep.py
@@ -1,0 +1,19 @@
+from __future__ import print_function
+from shutil import copyfile
+import os
+
+
+# Takes care of saving the files for the Neural Network Server.
+class ServerFiles():
+    def __init__(self, paths):
+        self.paths = paths
+
+    def _copy(self, f_name):
+        src_path = os.path.normpath(
+            self.paths.srvr_files_path + '/' + f_name)
+        dest_path = os.path.normpath(
+            self.paths.build_dir + '/' + f_name)
+        copyfile(src_path, dest_path)
+
+    def save_app(self):
+        self._copy('app.py')